+++ app/Http/Controllers/Backend/BrandController.php
@@ -135,9 +135,11 @@ class BrandController extends Controller
         $old_img = $request->old_image;
 
         if ($request->file('brand_image')) {
+
             if (file_exists($old_img)) {
                 unlink($old_img);
             }
+
             $image = $request->file('brand_image');
             $name_gen = hexdec(uniqid()) . '.' . $image->getClientOriginalExtension();
             Image::make($image)->resize(300, 300)->save('upload/brand/' . $name_gen);
@@ -186,7 +188,10 @@ class BrandController extends Controller
 
         $brand = Brand::findOrFail($id);
         $img = $brand->brand_image;
-        unlink($img);
+
+        if (file_exists($img)) {
+            unlink($img);
+        }
 
         Brand::findOrFail($id)->delete();
